Move audio, media and test_audio up a level
